Refactor root command
